Fixed Confirm on Power OFF and Reboot also minor changes
HyperVM did not ask for confirmation on reboot or power off as
the buttons were not created with updateForm, but only with update.
Added error messages and exceptions.

Added remaining "any" for IPv6 support.

Added status = 'on' when searching for VPS on Graph drawing as if
someone deleted the data directory then HyperVM failed to draw graphs.