Made iptables FORWARD consistent
Usinig and not usiing ACCEPT target in the forward chain was not consistent.
I decided to remove it from every iptables FORWARD chain rule as it would
overmatch other (user inserted) rules.
Also I replaced -A with -I FOWARD 1 so every matching rule is put in the
beginning of the chain.