Created new closeallinput.c
Honestly, Im not fully sure if this is all it's functionality, but tried to make it as good as possible (even tried to get more info using a decompiler, saw nothing else, though).
However! Now it become 32/64bit dependant, depending on compiling system. Hm. This is true to all sources.